Add size to linked list, refactor internals
This commit adds a size element to a linked list, and uses that in the
guard clause to get a specific element.

This also introduces a null object for an element is none is found, so
that it still responds to all the methods of an element and uses that
internally in the linked list.