Add to_a and reverse
This commit adds in converting a linked list to an array. It also
includes a naive implementation of a reverse method. This implementation
relies mostly on array functionality and should be updated at a future
date to be more performant or in line with what what should expect from
a linked list reversal. However, given the goal was playing with
minitest, the implementation was not a prime driver.